Refactor folder structure of the project. Move Validator and DockBlockReader into Validator folder
