+++ src/Validator/DockBlockReader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Valdator;
 
 class DockBlockReader
 {
